Refactoring. Move contact energy tables into a separate class, so that all the protein folders can use them. Also, introduce a wrapper function for energy table lookup, so that it is easy to change the energy table throughout the protein folder.
git-svn-id: http://wlab.ccbb.utexas.edu/svn/evoli@82 dadfaeae-b11a-0410-95f2-80a8e7ca9aa5

+++ trunk/test/t.protein.h
@@ -80,7 +80,7 @@ struct TEST_CLASS( protein_gene_basic )
 	{
 		CompactLatticeFolder* folder = new CompactLatticeFolder(side_length);
 		double max_dg = -5;
-		double sid = 574;
+		int sid = 574;
 		Gene g = GeneUtil::getSequenceForStructure(*folder, gene_length, max_dg, sid);
 		Protein p = g.translate();
 		FoldInfo fi = folder->fold(p);
